create western date allocation module
